requirements.txt
gitpython
openai
tqdm
python-dotenv
